+++ lib/src/api/restrr_builder.dart
@@ -3,6 +3,22 @@ import '../internal/requests/responses/rest_response.dart';
 import '../internal/restrr_impl.dart';
 import '../internal/utils/request_utils.dart';
 
+class SessionInfo {
+  final String name;
+  final String? description;
+  final SessionPlatform platform;
+
+  const SessionInfo({required this.name, this.description, required this.platform});
+
+  Map<String, dynamic> toJson() {
+    return {
+      'name': name,
+      if (description != null) 'description': description,
+      'platform': platform.name,
+    };
+  }
+}
+
 /// A builder for creating a new [Restrr] instance.
 /// The [Restrr] instance is created by calling [create].
 class RestrrBuilder {
@@ -19,32 +35,28 @@ class RestrrBuilder {
     return this;
   }
 
-  Future<Restrr> login({required String username, required String password, String? sessionName}) async {
+  Future<Restrr> login({required String username, required String password, required SessionInfo sessionInfo}) async {
     return _handleAuthProcess(authFunction: (apiImpl) {
       return apiImpl.requestHandler.apiRequest(
           route: SessionRoutes.create.compile(),
-          body: {
-            'username': username,
-            'password': password,
-            if (sessionName != null) 'session_name': sessionName,
-          },
+          body: {'username': username, 'password': password, ...sessionInfo.toJson()},
           noAuth: true,
           mapper: (json) => apiImpl.entityBuilder.buildPartialSession(json));
     });
   }
 
   Future<Restrr> register(
-      {required String username, required String password, String? displayName, String? email, String? sessionName}) async {
+      {required String username,
+      required String password,
+      required SessionInfo sessionInfo,
+      String? displayName,
+      String? email}) async {
     return _handleAuthProcess(authFunction: (apiImpl) async {
       // register user first
       final User user = await apiImpl.createUser(username: username, password: password, email: email);
       return apiImpl.requestHandler.apiRequest(
           route: SessionRoutes.create.compile(),
-          body: {
-            'username': user.username,
-            'password': password,
-            if (sessionName != null) 'session_name': sessionName,
-          },
+          body: {'username': user.username, 'password': password, ...sessionInfo.toJson()},
           noAuth: true,
           mapper: (json) => apiImpl.entityBuilder.buildPartialSession(json));
     });